+++ pkg/injector/patcher/sidecar_components_test.go
@@ -1,4 +1,17 @@
-package injector
+/*
+Copyright 2023 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package patcher
 
 import (
 	"encoding/json"
@@ -13,7 +26,6 @@ import (
 	commonapi "github.com/dapr/dapr/pkg/apis/common"
 	componentsapi "github.com/dapr/dapr/pkg/apis/components/v1alpha1"
 	"github.com/dapr/dapr/pkg/injector/annotations"
-	"github.com/dapr/dapr/pkg/injector/patcher"
 )
 
 func TestComponentsPatch(t *testing.T) {
@@ -62,12 +74,12 @@ func TestComponentsPatch(t *testing.T) {
 				},
 			},
 			jsonpatch.Patch{
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/env", []corev1.EnvVar{{
+				NewPatchOperation("add", PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
+				NewPatchOperation("add", PatchPathContainers+"/1/env", []corev1.EnvVar{{
 					Name:  ComponentsUDSMountPathEnvVar,
 					Value: socketSharedVolumeMount.MountPath,
 				}}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
+				NewPatchOperation("add", PatchPathContainers+"/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
 			},
 			&socketSharedVolumeMount,
 		},
@@ -92,12 +104,12 @@ func TestComponentsPatch(t *testing.T) {
 				},
 			},
 			jsonpatch.Patch{
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/env", []corev1.EnvVar{{
+				NewPatchOperation("add", PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
+				NewPatchOperation("add", PatchPathContainers+"/1/env", []corev1.EnvVar{{
 					Name:  ComponentsUDSMountPathEnvVar,
 					Value: socketSharedVolumeMount.MountPath,
 				}}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
+				NewPatchOperation("add", PatchPathContainers+"/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
 			},
 			&socketSharedVolumeMount,
 		},
@@ -130,25 +142,25 @@ func TestComponentsPatch(t *testing.T) {
 				},
 			},
 			jsonpatch.Patch{
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/env", []corev1.EnvVar{{
+				NewPatchOperation("add", PatchPathVolumes, []corev1.Volume{sharedComponentsSocketVolume()}),
+				NewPatchOperation("add", PatchPathContainers+"/1/env", []corev1.EnvVar{{
 					Name:  ComponentsUDSMountPathEnvVar,
 					Value: socketSharedVolumeMount.MountPath,
 				}}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes+"/-", corev1.Volume{
+				NewPatchOperation("add", PatchPathContainers+"/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
+				NewPatchOperation("add", PatchPathVolumes+"/-", corev1.Volume{
 					Name: "readonly",
 					VolumeSource: corev1.VolumeSource{
 						EmptyDir: &corev1.EmptyDirVolumeSource{},
 					},
 				}),
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes+"/-", corev1.Volume{
+				NewPatchOperation("add", PatchPathVolumes+"/-", corev1.Volume{
 					Name: "readwrite",
 					VolumeSource: corev1.VolumeSource{
 						EmptyDir: &corev1.EmptyDirVolumeSource{},
 					},
 				}),
-				patcher.NewPatchOperation("add", patcher.PatchPathContainers+"/-", corev1.Container{
+				NewPatchOperation("add", PatchPathContainers+"/-", corev1.Container{
 					Name:  componentName,
 					Image: componentImage,
 					Env: []corev1.EnvVar{
@@ -196,12 +208,12 @@ func TestComponentsPatch(t *testing.T) {
 				},
 			},
 			jsonpatch.Patch{
-				patcher.NewPatchOperation("add", patcher.PatchPathVolumes+"/-", []corev1.Volume{sharedComponentsSocketVolume()}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/env", []corev1.EnvVar{{
+				NewPatchOperation("add", PatchPathVolumes+"/-", []corev1.Volume{sharedComponentsSocketVolume()}),
+				NewPatchOperation("add", PatchPathContainers+"/1/env", []corev1.EnvVar{{
 					Name:  ComponentsUDSMountPathEnvVar,
 					Value: socketSharedVolumeMount.MountPath,
 				}}),
-				patcher.NewPatchOperation("add", "/spec/containers/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
+				NewPatchOperation("add", PatchPathContainers+"/1/volumeMounts", []corev1.VolumeMount{socketSharedVolumeMount}),
 			},
 			&socketSharedVolumeMount,
 		},
@@ -210,6 +222,7 @@ func TestComponentsPatch(t *testing.T) {
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
 			c := NewSidecarConfig(test.pod)
+			c.SetFromPodAnnotations()
 			_, componentContainers := c.SplitContainers()
 			patch, volumeMount := c.ComponentsPatchOps(componentContainers, Injectable(test.appID, test.componentsList))
 			patchJSON, _ := json.Marshal(patch)
