+++ pkg/injector/patcher/sidecar_container.go
@@ -1,4 +1,17 @@
-package injector
+/*
+Copyright 2023 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package patcher
 
 import (
 	"fmt"
@@ -212,7 +225,7 @@ func (cfg SidecarConfig) GetSidecarContainer(opts getSidecarContainerOpts) (*cor
 	// the Command and Args are passed as is. Otherwise, the Command is passed as a part of Args.
 	// This is to allow the Docker images to specify an ENTRYPOINT
 	// which is otherwise overridden by Command.
-	if podContainsTolerations(cfg.IgnoreEntrypointTolerations, cfg.Tolerations) {
+	if podContainsTolerations(cfg.IgnoreEntrypointTolerations, cfg.pod.Spec.Tolerations) {
 		container.Command = cmd
 		container.Args = args
 	} else {
